requirements.txt
streamlit>=1.20.0
requests>=2.28.0
fastapi>=0.118.0
uvicorn>=0.37.0
langchain>=0.3.27
langchain-community>=0.3.30
langchain-groq>=0.2.0
groq>=0.10.0
chromadb>=1.1.0
pydantic>=2.11.9
python-multipart>=0.0.20
rank-bm25>=0.2.2
python-dotenv>=1.1.1
pdfplumber>=0.11.7
python-docx>=1.2.0